Move generateTransactionExportFileName() to src/renderer/utils/transactionFile.ts.`Move generateTransactionExportContent() to src/renderer/utils/transactionFile.ts and rename to generateTransactionV1ExportContent().
